In this project we will find locations of ambulance stations that minimize the maximum response time to any location on a given map under a given budget.
We will solve the problem into four steps. The first step to solve the problem is to describe the problem, formulate the objective function and constraints along with the introduction of variables. The next step is to reformulate the problem so we can solve it using MATLAB. In the third step we will work with the most interesting part which is to write the matrices in the MATLAB program. The last step is to solve a small example.